requirements.txt
# Core dependencies
uiautomator2==2.16.23
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.5
pydantic-settings==2.7.0

# Async task queue
celery==5.4.0
redis==5.2.1

# Database (已切换到MCP，以下为备用)
# SQLAlchemy==2.0.23
# alembic==1.12.1
# psycopg2-binary==2.9.9

# HTTP Client for MCP
httpx==0.28.1
aiohttp==3.11.11

# Logging
loguru==0.7.2

# Retry mechanism
tenacity==8.2.3

# Additional utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# AI models
openai==1.59.7
google-generativeai==0.8.3

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
pytest-mock==3.14.0
